Mark pkcs7_supply_detached_data as an exported symbol so it can be accessed from modules.
This change will not be necessary in the final project, because our changes will not exist in a kernel module but rather the core kernel image. That being said, I am currently investigating the use of the PKCS#7 APIs in a kernel module to see if they will be suitable for use in our project, which is why this symbol is being exported.